+++ courses/models.py
@@ -4,6 +4,8 @@ from django.conf import settings
 from django.utils import timezone
 import datetime
 
+from users.models import User
+
 
 class Course(models.Model):
     CATEGORY_CHOICES = [
@@ -218,3 +220,20 @@ class Assignment(models.Model):
 
     def __str__(self):
         return f"{self.course.title} - Assignment: {self.title}"
+    
+
+class Notification(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="notifications")
+    message = models.CharField(max_length=255)
+    url = models.CharField(max_length=255, blank=True, null=True)  
+    created_at = models.DateTimeField(default=timezone.now)
+    is_read = models.BooleanField(default=False)
+
+    def __str__(self):
+        return f"{self.user.username} - {self.message[:30]}"
+    
+class LiveClassAttendance(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    live_class = models.ForeignKey(LiveClass, on_delete=models.CASCADE)
+    joined_at = models.DateTimeField(null=True, blank=True)
+    duration = models.IntegerField(default=0)
